Make halide-arrayfire compile again.
Updates the flake to point to the latest arrayfire-nix and
arrayfire-haskell such that halide-arrayfire can be compiled again.

Adds ShapedArray class such that more than four-dimensional arrays can
be passed to withHalideBuffer. mkShapedArray and produceShapedArray are
helper functions to ease the use of ShapedArrays.

Finally, fixes some compilation errors in the tutorials.